V2 Phase 1 (fondations) : montants en Decimal, navigation restructuree, audit securite
Suite au nouveau cahier des charges V2 : premiere des 9 phases prevues,
centree sur les fondations avant d'attaquer les nouveaux modules metier.

1. Montants en Decimal, plus jamais en Float (backend/app/models.py)
   Toutes les colonnes monetaires (quantite, prix_unitaire_ht, taux_tva,
   acompte_pourcentage, montant, budget) passent de Float a Numeric(10,2).
   Float accumule des erreurs d'arrondi binaire au fil des lectures/
   ecritures en base ; Numeric stocke la valeur exacte. Corrige au passage
   deux endroits qui melangeaient encore un Decimal avec un float litteral
   (defaultdict(float) dans analytics.py, un "or 0.0" dans pdf.py) : ces
   melanges levent TypeError en Python, trouves en testant reellement
   chaque endpoint apres la migration plutot qu'en la supposant correcte.
   Les schemas Pydantic restent en float (Pydantic v2 convertit Decimal en
   float automatiquement) : le contrat API et le frontend ne changent pas.

2. Navigation restructuree selon le cahier des charges (section 4)
   - "Tableau de bord" -> "Accueil"
   - "Clients & prospects" separe en deux : "Prospects" (pipeline complet,
     Kanban inchange) et "Clients" (nouveau : annuaire des affaires
     gagnees, plus simple qu'un Kanban puisqu'il n'y a plus de statut a
     faire avancer)
   - "Conformite" absorbee dans une nouvelle section "Entreprise", qui
     ajoute aussi un vrai formulaire d'edition des informations de
     l'entreprise (SIRET, assurance, coordonnees) qui n'existait nulle
     part avant (seulement modifiable a l'inscription jusqu'ici)
   - Nouvelle section "Statistiques" branchee sur le backend analytics.py
     qui existait deja mais n'avait aucune interface

3. Audit de securite rapide (backend/app/main.py, config.py)
   - CORS_ORIGINS devient configurable (defaut "*" pour ne rien casser en
     dev), au lieu d'un wildcard code en dur
   - Refus de demarrer si JWT_SECRET garde sa valeur de developpement ET
     que DATABASE_URL pointe vers une vraie base (Postgres) plutot que
     SQLite - simple avertissement en dev, erreur bloquante si ca
     ressemble a un vrai déploiement
   - Reste a traiter plus tard (necessite des decisions d'infrastructure) :
     rate limiting, upload de fichiers reel pour les documents.

Teste avec Playwright (Chromium) + un compte de demonstration realiste :
migration Decimal verifiee sur tous les endpoints qui calculent des
montants (devis, factures, chantiers, dashboard, analytics, PDF,
duplication de devis) apres correction des deux bugs trouves ; navigation
verifiee ecran par ecran (Prospects garde ses 9 colonnes, Clients affiche
les 2 clients "gagne" de la demo, formulaire Entreprise sauvegarde et
persiste apres rechargement, Statistiques affiche les vraies donnees) ;
isolation multi-tenant reverifiee sur les nouveaux endpoints avec un
deuxieme compte (clients, devis, PDF, analytics tous correctement vides/
404/402 pour un compte qui n'a pas les donnees ou l'abonnement).

+++ backend/app/routers/analytics.py
@@ -1,5 +1,6 @@
 from collections import defaultdict
 from datetime import date, timedelta
+from decimal import Decimal
 
 from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session, joinedload
@@ -30,7 +31,7 @@ def obtenir_analytics(
         .filter(Facture.artisan_id == artisan.id, Paiement.date_paiement >= depuis_12_mois)
         .all()
     )
-    ca_par_mois_dict = defaultdict(float)
+    ca_par_mois_dict = defaultdict(Decimal)
     for date_paiement, montant in paiements:
         cle = date_paiement.strftime("%Y-%m")
         ca_par_mois_dict[cle] += montant
